PythonLightStar/pyproject.toml
[project]
name = "PythonLightStar"
version = "0.1.0"
description = "Performance focused Litestar example with Granian and Psycopg"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "granian[uvloop]>=2.5.6",
    "litestar>=2.18.0",
    "litestar-psycopg>=0.1.1",
    "msgspec>=0.19.0",
    "psycopg[binary,pool]>=3.2.12",
    "python-dotenv>=1.0.0",
    "uvloop>=0.22.1",
]

[tool.localenv]
pythonVersion = "3.14.0"
venv = "3.14.0"
venvPath = "/Users/admin/.pyenv/versions"

[tool.pyright]
pythonVersion = "3.14"
venv = "3.14.0"
venvPath = "/Users/admin/.pyenv/versions"
reportAttributeAccessIssue = "none"
